// pos/src/app/models/m-style-master.ts
import { MItemImageMaster } from './m-item-image-master';
import { MColorMaster } from './m-color-master';

export class MStyleMaster {
    constructor() { }
    id?: number;
    styleCode?: string;
    styleName?: string;
    shortDesignName?: string;
    styleSegmentation?: number;
    itemImage?: string;
    remarks?: string;
    designID?: number;
    designName?: string;
    productDepartmentCode?: string;
    brandID?: number;
    subBrandID?: number;
    collectionID?: number;
    armadaCollectionID?: number;
    divisionID?: number;
    productGroupID?: number;
    productSubGroupID?: number;
    seasonID?: number;
    yearCode?: number;
    productLineID?: number;
    styleStatusID?: number;
    designerID?: number;
    purchasePriceListID?: number;
    purchasePrice?: number;
    salesPriceListID?: number;
    salesPrice?: number;
    purchaseCurrencyID?: number;
    rrpPrice?: number;
    rrpCurrencyID?: number;
    documentID?: number;
    scaleID?: number;
    franchise?: boolean;
    exchangeRate?: number;
    countryID?: number;
    storeID?: number;
    stateID?: number;
    sizeID?: number;
    sizeName?: string;
    sizeCode?: string;
    visualOrder?: number;
    colorID?: number;
    colorCode?: string;
    colorName?: string;
    scaleDetailMasterList?: Array<MStyleMaster>;
    colorMasterList?: Array<MColorMaster>;
    itemImageMasterList?: Array<MItemImageMaster>;
    styleRunningNum?: number;
    salesType?: string;
    returnIDs?: string;
    composition?: string;
    symbolGroup?: string;
    owner?: string;
    countryOfOrigin?: string;
    shortDescriptionn?: string;
    dropID?: number;
    grade?: string;
    developmentOffice?: string;
    supplierBarcode?: string;
    arabicStyle?: string;
    arabicSKU?: string;
    importExcelList?: Array<MStyleMaster>;
    importcolorExcelList?: Array<MStyleMaster>;
    importScaleExcelList?: Array<MStyleMaster>;
    seasonCode?: string;
    seasonName?: string;
    active?: boolean;
    brandCode?:string;    
    subBrandCode?:string;    
    collectionCode?:string;    
    armadaCollectionCode?:string;    
    divisionCode?:string;    
    productGroupCode?:string;    
    productLineCode?:string;    
    styleStatusCode?:string;    
    designerCode?:string;    
    purchasePriceListCode?:string;    
    purchaseCurrencyCode?:string;    
    rrpCurrencyCode?:string;    
    scaleCode?:string;    
    segmentationCode?:string;    
    designCode?:string;    
    productSubGroupCode?:string;
}